requirements.txt
# VoiceForge-Nextgen Dependencies
# Python 3.12.9 Required

# Core Audio Processing
pyaudio==0.2.14
sounddevice==0.4.6
numpy==1.26.4
scipy==1.11.4

# AI/ML - RVC Dependencies
# Note: PyTorch 2.1.2 not available for Python 3.12, using latest compatible version
torch>=2.2.0
torchaudio>=2.2.0
torchvision>=0.17.0
fairseq==0.12.2
librosa==0.10.1
faiss-cpu==1.7.4  # Use faiss-gpu if CUDA available
praat-parselmouth==0.4.3
pyworld==0.3.4
resampy==0.4.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
colorlog==6.8.0
psutil==5.9.7

# GUI
PyQt6==6.6.1
PyQt6-Qt6==6.6.1

# Development/Debug
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
mypy==1.8.0

# Optional Performance (uncomment if needed)
# onnxruntime-gpu==1.16.3
# tensorrt==8.6.1